requirements.txt
pyspark==3.3.1
psycopg2
pytest
python-dotenv
minio